Crate rome_js_syntax

A crate for generated Syntax node definitions and utility macros. Both rome_js_lexer and rome_js_parser rely on these definitions, therefore they are wrapped in this crate to prevent cyclic dependencies

Re-exports

pub use expr_ext::*;

pub use modifier_ext::*;

pub use source_type::*;

pub use stmt_ext::*;

Modules

expr_ext

Extensions for things which are not easily generated in ast expr nodes

import_ext

jsx_ext

kind

Generated file, do not edit by hand, see xtask/codegen

macros

Generated file, do not edit by hand, see xtask/codegen

modifier_ext

numbers

JS Number parsing.

source_type

stmt_ext

Extended AST node definitions for statements which are unique and special enough to generate code for manually

suppression

Macros

T

Utility macro for creating a SyntaxKind through simple macro syntax

map_syntax_node

Reconstruct an AstNode from a SyntaxNode

Structs

DebugSyntaxElementChildren

ImportMeta

ImportMetaFields

JsArrayAssignmentPattern

JsArrayAssignmentPatternElementList

JsArrayAssignmentPatternFields

JsArrayAssignmentPatternRestElement

JsArrayAssignmentPatternRestElementFields

JsArrayBindingPattern

JsArrayBindingPatternElementList

JsArrayBindingPatternFields

JsArrayBindingPatternRestElement

JsArrayBindingPatternRestElementFields

JsArrayElementList

JsArrayExpression

JsArrayExpressionFields

JsArrayHole

JsArrayHoleFields

JsArrowFunctionExpression

JsArrowFunctionExpressionFields

JsAssignmentExpression

JsAssignmentExpressionFields

JsAssignmentWithDefault

JsAssignmentWithDefaultFields

JsAwaitExpression

JsAwaitExpressionFields

JsBigIntLiteralExpression

JsBigIntLiteralExpressionFields

JsBinaryExpression

JsBinaryExpressionFields

JsBindingPatternWithDefault

JsBindingPatternWithDefaultFields

JsBlockStatement

JsBlockStatementFields

JsBooleanLiteralExpression

JsBooleanLiteralExpressionFields

JsBreakStatement

JsBreakStatementFields

JsCallArgumentList

JsCallArguments

JsCallArgumentsFields

JsCallExpression

JsCallExpressionFields

JsCaseClause

JsCaseClauseFields

JsCatchClause

JsCatchClauseFields

JsCatchDeclaration

JsCatchDeclarationFields

JsClassDeclaration

JsClassDeclarationFields

JsClassExportDefaultDeclaration

JsClassExportDefaultDeclarationFields

JsClassExpression

JsClassExpressionFields

JsClassMemberList

JsComputedMemberAssignment

JsComputedMemberAssignmentFields

JsComputedMemberExpression

JsComputedMemberExpressionFields

JsComputedMemberName

JsComputedMemberNameFields

JsConditionalExpression

JsConditionalExpressionFields

JsConstructorClassMember

JsConstructorClassMemberFields

JsConstructorModifierList

JsConstructorParameterList

JsConstructorParameters

JsConstructorParametersFields

JsContinueStatement

JsContinueStatementFields

JsDebuggerStatement

JsDebuggerStatementFields

JsDefaultClause

JsDefaultClauseFields

JsDefaultImportSpecifier

JsDefaultImportSpecifierFields

JsDirective

JsDirectiveFields

JsDirectiveList

JsDoWhileStatement

JsDoWhileStatementFields

JsElseClause

JsElseClauseFields

JsEmptyClassMember

JsEmptyClassMemberFields

JsEmptyStatement

JsEmptyStatementFields

JsExport

JsExportAsClause

JsExportAsClauseFields

JsExportDefaultDeclarationClause

JsExportDefaultDeclarationClauseFields

JsExportDefaultExpressionClause

JsExportDefaultExpressionClauseFields

JsExportFields

JsExportFromClause

JsExportFromClauseFields

JsExportNamedClause

JsExportNamedClauseFields

JsExportNamedFromClause

JsExportNamedFromClauseFields

JsExportNamedFromSpecifier

JsExportNamedFromSpecifierFields

JsExportNamedFromSpecifierList

JsExportNamedShorthandSpecifier

JsExportNamedShorthandSpecifierFields

JsExportNamedSpecifier

JsExportNamedSpecifierFields

JsExportNamedSpecifierList

JsExpressionSnipped

JsExpressionSnippedFields

JsExpressionStatement

JsExpressionStatementFields

JsExtendsClause

JsExtendsClauseFields

JsFinallyClause

JsFinallyClauseFields

JsForInStatement

JsForInStatementFields

JsForOfStatement

JsForOfStatementFields

JsForStatement

JsForStatementFields

JsForVariableDeclaration

JsForVariableDeclarationFields

JsFormalParameter

JsFormalParameterFields

JsFunctionBody

JsFunctionBodyFields

JsFunctionDeclaration

JsFunctionDeclarationFields

JsFunctionExportDefaultDeclaration

JsFunctionExportDefaultDeclarationFields

JsFunctionExpression

JsFunctionExpressionFields

JsGetterClassMember

JsGetterClassMemberFields

JsGetterObjectMember

JsGetterObjectMemberFields

JsIdentifierAssignment

JsIdentifierAssignmentFields

JsIdentifierBinding

JsIdentifierBindingFields

JsIdentifierExpression

JsIdentifierExpressionFields

JsIfStatement

JsIfStatementFields

JsImport

JsImportAssertion

JsImportAssertionEntry

JsImportAssertionEntryFields

JsImportAssertionEntryList

JsImportAssertionFields

JsImportBareClause

JsImportBareClauseFields

JsImportCallExpression

JsImportCallExpressionFields

JsImportDefaultClause

JsImportDefaultClauseFields

JsImportFields

JsImportNamedClause

JsImportNamedClauseFields

JsImportNamespaceClause

JsImportNamespaceClauseFields

JsInExpression

JsInExpressionFields

JsInitializerClause

JsInitializerClauseFields

JsInstanceofExpression

JsInstanceofExpressionFields

JsLabeledStatement

JsLabeledStatementFields

JsLanguage

JsLiteralExportName

JsLiteralExportNameFields

JsLiteralMemberName

JsLiteralMemberNameFields

JsLogicalExpression

JsLogicalExpressionFields

JsMethodClassMember

JsMethodClassMemberFields

JsMethodModifierList

JsMethodObjectMember

JsMethodObjectMemberFields

JsModule

JsModuleFields

JsModuleItemList

JsModuleSource

JsModuleSourceFields

JsName

JsNameFields

JsNamedImportSpecifier

JsNamedImportSpecifierFields

JsNamedImportSpecifierList

JsNamedImportSpecifiers

JsNamedImportSpecifiersFields

JsNamespaceImportSpecifier

JsNamespaceImportSpecifierFields

JsNewExpression

JsNewExpressionFields

JsNullLiteralExpression

JsNullLiteralExpressionFields

JsNumberLiteralExpression

JsNumberLiteralExpressionFields

JsObjectAssignmentPattern

JsObjectAssignmentPatternFields

JsObjectAssignmentPatternProperty

JsObjectAssignmentPatternPropertyFields

JsObjectAssignmentPatternPropertyList

JsObjectAssignmentPatternRest

JsObjectAssignmentPatternRestFields

JsObjectAssignmentPatternShorthandProperty

JsObjectAssignmentPatternShorthandPropertyFields

JsObjectBindingPattern

JsObjectBindingPatternFields

JsObjectBindingPatternProperty

JsObjectBindingPatternPropertyFields

JsObjectBindingPatternPropertyList

JsObjectBindingPatternRest

JsObjectBindingPatternRestFields

JsObjectBindingPatternShorthandProperty

JsObjectBindingPatternShorthandPropertyFields

JsObjectExpression

JsObjectExpressionFields

JsObjectMemberList

JsParameterList

JsParameters

JsParametersFields

JsParenthesizedAssignment

JsParenthesizedAssignmentFields

JsParenthesizedExpression

JsParenthesizedExpressionFields

JsPostUpdateExpression

JsPostUpdateExpressionFields

JsPreUpdateExpression

JsPreUpdateExpressionFields

JsPrivateClassMemberName

JsPrivateClassMemberNameFields

JsPrivateName

JsPrivateNameFields

JsPropertyClassMember

JsPropertyClassMemberFields

JsPropertyModifierList

JsPropertyObjectMember

JsPropertyObjectMemberFields

JsReferenceIdentifier

JsReferenceIdentifierFields

JsRegexLiteralExpression

JsRegexLiteralExpressionFields

JsRestParameter

JsRestParameterFields

JsReturnStatement

JsReturnStatementFields

JsScript

JsScriptFields

JsSequenceExpression

JsSequenceExpressionFields

JsSetterClassMember

JsSetterClassMemberFields

JsSetterObjectMember

JsSetterObjectMemberFields

JsShorthandNamedImportSpecifier

JsShorthandNamedImportSpecifierFields

JsShorthandPropertyObjectMember

JsShorthandPropertyObjectMemberFields

JsSpread

JsSpreadFields

JsStatementList

JsStaticInitializationBlockClassMember

JsStaticInitializationBlockClassMemberFields

JsStaticMemberAssignment

JsStaticMemberAssignmentFields

JsStaticMemberExpression

JsStaticMemberExpressionFields

JsStaticModifier

JsStaticModifierFields

JsStringLiteralExpression

JsStringLiteralExpressionFields

JsSuperExpression

JsSuperExpressionFields

JsSwitchCaseList

JsSwitchStatement

JsSwitchStatementFields

JsTemplate

JsTemplateChunkElement

JsTemplateChunkElementFields

JsTemplateElement

JsTemplateElementFields

JsTemplateElementList

JsTemplateFields

JsThisExpression

JsThisExpressionFields

JsThrowStatement

JsThrowStatementFields

JsTryFinallyStatement

JsTryFinallyStatementFields

JsTryStatement

JsTryStatementFields

JsUnaryExpression

JsUnaryExpressionFields

JsUnknown

JsUnknownAssignment

JsUnknownBinding

JsUnknownExpression

JsUnknownImportAssertionEntry

JsUnknownMember

JsUnknownNamedImportSpecifier

JsUnknownParameter

JsUnknownStatement

JsVariableDeclaration

JsVariableDeclarationClause

JsVariableDeclarationClauseFields

JsVariableDeclarationFields

JsVariableDeclarator

JsVariableDeclaratorFields

JsVariableDeclaratorList

JsVariableStatement

JsVariableStatementFields

JsWhileStatement

JsWhileStatementFields

JsWithStatement

JsWithStatementFields

JsYieldArgument

JsYieldArgumentFields

JsYieldExpression

JsYieldExpressionFields

JsxAttribute

JsxAttributeFields

JsxAttributeInitializerClause

JsxAttributeInitializerClauseFields

JsxAttributeList

JsxChildList

JsxClosingElement

JsxClosingElementFields

JsxClosingFragment

JsxClosingFragmentFields

JsxElement

JsxElementFields

JsxExpressionAttributeValue

JsxExpressionAttributeValueFields

JsxExpressionChild

JsxExpressionChildFields

JsxFragment

JsxFragmentFields

JsxMemberName

JsxMemberNameFields

JsxName

JsxNameFields

JsxNamespaceName

JsxNamespaceNameFields

JsxOpeningElement

JsxOpeningElementFields

JsxOpeningFragment

JsxOpeningFragmentFields

JsxReferenceIdentifier

JsxReferenceIdentifierFields

JsxSelfClosingElement

JsxSelfClosingElementFields

JsxSpreadAttribute

JsxSpreadAttributeFields

JsxSpreadChild

JsxSpreadChildFields

JsxString

JsxStringFields

JsxTagExpression

JsxTagExpressionFields

JsxText

JsxTextFields

NewTarget

NewTargetFields

SyntaxNodeText

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

TsAbstractModifier

TsAbstractModifierFields

TsAccessibilityModifier

TsAccessibilityModifierFields

TsAnyType

TsAnyTypeFields

TsArrayType

TsArrayTypeFields

TsAsAssignment

TsAsAssignmentFields

TsAsExpression

TsAsExpressionFields

TsAssertsCondition

TsAssertsConditionFields

TsAssertsReturnType

TsAssertsReturnTypeFields

TsBigIntLiteralType

TsBigIntLiteralTypeFields

TsBigintType

TsBigintTypeFields

TsBooleanLiteralType

TsBooleanLiteralTypeFields

TsBooleanType

TsBooleanTypeFields

TsCallSignatureTypeMember

TsCallSignatureTypeMemberFields

TsConditionalType

TsConditionalTypeFields

TsConstructSignatureTypeMember

TsConstructSignatureTypeMemberFields

TsConstructorSignatureClassMember

TsConstructorSignatureClassMemberFields

TsConstructorType

TsConstructorTypeFields

TsDeclareFunctionDeclaration

TsDeclareFunctionDeclarationFields

TsDeclareModifier

TsDeclareModifierFields

TsDeclareStatement

TsDeclareStatementFields

TsDefaultTypeClause

TsDefaultTypeClauseFields

TsDefinitePropertyAnnotation

TsDefinitePropertyAnnotationFields

TsDefiniteVariableAnnotation

TsDefiniteVariableAnnotationFields

TsEmptyExternalModuleDeclarationBody

TsEmptyExternalModuleDeclarationBodyFields

TsEnumDeclaration

TsEnumDeclarationFields

TsEnumMember

TsEnumMemberFields

TsEnumMemberList

TsExportAsNamespaceClause

TsExportAsNamespaceClauseFields

TsExportAssignmentClause

TsExportAssignmentClauseFields

TsExportDeclareClause

TsExportDeclareClauseFields

TsExtendsClause

TsExtendsClauseFields

TsExternalModuleDeclaration

TsExternalModuleDeclarationFields

TsExternalModuleReference

TsExternalModuleReferenceFields

TsFunctionType

TsFunctionTypeFields

TsGetterSignatureClassMember

TsGetterSignatureClassMemberFields

TsGetterSignatureTypeMember

TsGetterSignatureTypeMemberFields

TsGlobalDeclaration

TsGlobalDeclarationFields

TsIdentifierBinding

TsIdentifierBindingFields

TsImplementsClause

TsImplementsClauseFields

TsImportEqualsDeclaration

TsImportEqualsDeclarationFields

TsImportType

TsImportTypeFields

TsImportTypeQualifier

TsImportTypeQualifierFields

TsIndexSignatureClassMember

TsIndexSignatureClassMemberFields

TsIndexSignatureModifierList

TsIndexSignatureParameter

TsIndexSignatureParameterFields

TsIndexSignatureTypeMember

TsIndexSignatureTypeMemberFields

TsIndexedAccessType

TsIndexedAccessTypeFields

TsInferType

TsInferTypeFields

TsInstantiationExpression

TsInstantiationExpressionFields

TsInterfaceDeclaration

TsInterfaceDeclarationFields

TsIntersectionType

TsIntersectionTypeElementList

TsIntersectionTypeFields

TsMappedType

TsMappedTypeAsClause

TsMappedTypeAsClauseFields

TsMappedTypeFields

TsMappedTypeOptionalModifierClause

TsMappedTypeOptionalModifierClauseFields

TsMappedTypeReadonlyModifierClause

TsMappedTypeReadonlyModifierClauseFields

TsMethodSignatureClassMember

TsMethodSignatureClassMemberFields

TsMethodSignatureModifierList

TsMethodSignatureTypeMember

TsMethodSignatureTypeMemberFields

TsModuleBlock

TsModuleBlockFields

TsModuleDeclaration

TsModuleDeclarationFields

TsNameWithTypeArguments

TsNameWithTypeArgumentsFields

TsNamedTupleTypeElement

TsNamedTupleTypeElementFields

TsNeverType

TsNeverTypeFields

TsNonNullAssertionAssignment

TsNonNullAssertionAssignmentFields

TsNonNullAssertionExpression

TsNonNullAssertionExpressionFields

TsNonPrimitiveType

TsNonPrimitiveTypeFields

TsNullLiteralType

TsNullLiteralTypeFields

TsNumberLiteralType

TsNumberLiteralTypeFields

TsNumberType

TsNumberTypeFields

TsObjectType

TsObjectTypeFields

TsOptionalPropertyAnnotation

TsOptionalPropertyAnnotationFields

TsOptionalTupleTypeElement

TsOptionalTupleTypeElementFields

TsOverrideModifier

TsOverrideModifierFields

TsParenthesizedType

TsParenthesizedTypeFields

TsPredicateReturnType

TsPredicateReturnTypeFields

TsPropertyParameter

TsPropertyParameterFields

TsPropertyParameterModifierList

TsPropertySignatureClassMember

TsPropertySignatureClassMemberFields

TsPropertySignatureModifierList

TsPropertySignatureTypeMember

TsPropertySignatureTypeMemberFields

TsQualifiedModuleName

TsQualifiedModuleNameFields

TsQualifiedName

TsQualifiedNameFields

TsReadonlyModifier

TsReadonlyModifierFields

TsReferenceType

TsReferenceTypeFields

TsRestTupleTypeElement

TsRestTupleTypeElementFields

TsReturnTypeAnnotation

TsReturnTypeAnnotationFields

TsSetterSignatureClassMember

TsSetterSignatureClassMemberFields

TsSetterSignatureTypeMember

TsSetterSignatureTypeMemberFields

TsStringLiteralType

TsStringLiteralTypeFields

TsStringType

TsStringTypeFields

TsSymbolType

TsSymbolTypeFields

TsTemplateChunkElement

TsTemplateChunkElementFields

TsTemplateElement

TsTemplateElementFields

TsTemplateElementList

TsTemplateLiteralType

TsTemplateLiteralTypeFields

TsThisParameter

TsThisParameterFields

TsThisType

TsThisTypeFields

TsTupleType

TsTupleTypeElementList

TsTupleTypeFields

TsTypeAliasDeclaration

TsTypeAliasDeclarationFields

TsTypeAnnotation

TsTypeAnnotationFields

TsTypeArgumentList

TsTypeArguments

TsTypeArgumentsFields

TsTypeAssertionAssignment

TsTypeAssertionAssignmentFields

TsTypeAssertionExpression

TsTypeAssertionExpressionFields

TsTypeConstraintClause

TsTypeConstraintClauseFields

TsTypeList

TsTypeMemberList

TsTypeOperatorType

TsTypeOperatorTypeFields

TsTypeParameter

TsTypeParameterFields

TsTypeParameterList

TsTypeParameterName

TsTypeParameterNameFields

TsTypeParameters

TsTypeParametersFields

TsTypeofType

TsTypeofTypeFields

TsUndefinedType

TsUndefinedTypeFields

TsUnionType

TsUnionTypeFields

TsUnionTypeVariantList

TsUnknownType

TsUnknownTypeFields

TsVoidType

TsVoidTypeFields

Enums

JsAnyArrayAssignmentPatternElement

JsAnyArrayBindingPatternElement

JsAnyArrayElement

JsAnyArrowFunctionParameters

JsAnyAssignment

JsAnyAssignmentPattern

JsAnyBinding

JsAnyBindingPattern

JsAnyCallArgument

JsAnyClass

JsAnyClassMember

JsAnyClassMemberName

JsAnyConstructorParameter

JsAnyDeclaration

JsAnyDeclarationClause

JsAnyExportClause

JsAnyExportDefaultDeclaration

JsAnyExportNamedSpecifier

JsAnyExpression

JsAnyForInOrOfInitializer

JsAnyForInitializer

JsAnyFormalParameter

JsAnyFunction

JsAnyFunctionBody

JsAnyImportAssertionEntry

JsAnyImportClause

JsAnyInProperty

JsAnyLiteralExpression

JsAnyMethodModifier

JsAnyModuleItem

JsAnyName

JsAnyNamedImport

JsAnyNamedImportSpecifier

JsAnyObjectAssignmentPatternMember

JsAnyObjectBindingPatternMember

JsAnyObjectMember

JsAnyObjectMemberName

JsAnyParameter

JsAnyPropertyModifier

JsAnyRoot

JsAnyStatement

JsAnySwitchClause

JsAnyTemplateElement

JsSyntaxKind

The kind of syntax node, e.g. IDENT, FUNCTION_KW, or FOR_STMT.

JsxAnyAttribute

JsxAnyAttributeName

JsxAnyAttributeValue

JsxAnyChild

JsxAnyElementName

JsxAnyName

JsxAnyObjectName

JsxAnyTag

OperatorPrecedence

See: MDN Operator precedence

TokenAtOffset

There might be zero, one or two leaves at a given offset.

TriviaPieceKind

TsAnyExternalModuleDeclarationBody

TsAnyIndexSignatureModifier

TsAnyMethodSignatureModifier

TsAnyModuleName

TsAnyModuleReference

TsAnyName

TsAnyPropertyAnnotation

TsAnyPropertyParameterModifier

TsAnyPropertySignatureAnnotation

TsAnyPropertySignatureModifier

TsAnyReturnType

TsAnyTemplateElement

TsAnyTupleTypeElement

TsAnyTypeMember

TsAnyTypePredicateParameterName

TsAnyVariableAnnotation

TsType

WalkEvent

WalkEvent describes tree walking process.

Traits

TextLen

Primitives with a textual length that can be passed to TextSize::of.

Type Definitions

JsSyntaxElement

JsSyntaxElementChildren

JsSyntaxList

JsSyntaxNode

JsSyntaxNodeChildren

JsSyntaxToken

JsSyntaxTrivia